refactor(workstation): merge duplicative sections + refine hero
The page ran five demo/showcase sections back to back, two of which were the
*same* interactive list+preview component (views and cross-view workflows),
and a thin one-GIF "search" section whose navigation GIF duplicated what the
view explorer already shows.

- Merge the views explorer and cross-view showcase into one WorkstationShowcase
  with a segmented toggle (16 views ⇄ cross-view actions) over a shared
  list+preview. Deletes ViewExplorer + CrossViewShowcase.
- Fold the standalone "search" section into the navigation section as
  "Keyboard-first, mouse-free" (chord jump + / filter), and swap the
  redundant view-switching GIF for the unique live-filter demo.
- Net: nine sections down to seven, no duplicated patterns.
- Refine the hero copy to lead with the pain (juggling GUI / terminal /
  browser) and the concrete surfaces, sharpened by AI.
